New commands: sgml-split-element (C-c RET): Split the current element at point.  If repeated, the containing element will be split before the beginning of then current element (and then the element containing the containing element).
sgml-make-character-reference (C-c #): Convert a character to a
character reference.  Convert a character reference to a character
with C-u C-c #.

Attribute editing

If the point is immediately after a start-tag, sgml-edit-attrib will
edit the attributes of that tag, otherwise the current elements
attributes will be editied.  The reason for this change is to make it
easier to edit the attributes of an inserted empty element.

There is now a command sgml-insert-attribute that asks for a attribute
name and value and inserts/changes this attribute.  The element
affected is the same as for sgml-edit-attrib.

Attributes can also be inserted from a menu.  The menu is available
from the menu bar in "Markup".  Alternatively you can put the point
inside the start-tag and use S-mouse-1 (for lucid it is mouse-3).

This menu will have one pane for every attribute that has a token list
(or notation list) as declared valued, and one pane for all the other
attributes.

(sgml-normalize-start-tag): use sgml-change-start-tag
(sgml-make-character-reference): new command
(sgml-split-element): New command
(sgml-options-menu): uses sgml-user-options in a general way.
(sgml-insert-start-tag): new function
(sgml-change-start-tag): new function
(sgml-find-attribute-element): new function
(sgml-attrib-menu): new function
(sgml-tags-menu): pop up a attribute menu if called when point is
  inside a start-tag.
(sgml-element-attval): new function.
(sgml-set-live-element-indicator): bind deactivate-mark.
(sgml-find-context-of): don't change point.
(sgml-set-parse-state): fix bug that that set markup type of start tag
  for empty element to end-tag.
(sgml-parse-minimum-literal): New function
(sgml-parse-external): use parse-minimum-literal for public
  identifier.
(sgml-formal-pubid-regexp): simplify, now that parse-minimum-literal
  does the right thing.
(sgml-pubid-parts): %% => %
(sgml-find-previous-element): takes optional argument: IN-ELEMENT, the
  enclosing element.
(sgml-backward-element): use sgml-last-element
(sgml-find-element-after): take IN-ELEMENT argument
(sgml-backward-element): use sgml-last-element
(sgml-beginning-of-element): use sgml-last-element
(sgml-end-of-element): use sgml-last-element
(sgml-down-element): use sgml-last-element
(sgml-users-of-last-element): add new commancds